refactor: introduce infrastructure layer
External utilities are moved to a separate directory called
`infrastructure`. This includes logging and terminal functions.
Some of which were contained in a `utils.rs` file. The contents of that
file are now neatly sorted into either `terminal.rs`, `logging.rs` or
`macros.rs`
