+++ src/app.rs
@@ -1,4 +1,6 @@
+use crate::infrastructure::logging::Logger;
 use crate::{
+    infrastructure::logging,
     log_on_error,
     ui::popup::{info_popup::InfoPopUp, PopUp},
 };
@@ -6,7 +8,6 @@ use ansi_to_tui::IntoText;
 use color_eyre::eyre::bail;
 use config::Config;
 use cover_renderer::render_cover;
-use logging::Logger;
 use patch_hub::lore::{
     lore_api_client::BlockingLoreAPIClient,
     lore_session,
@@ -24,11 +25,8 @@ use screens::{
 };
 use std::collections::{HashMap, HashSet};
 
-use crate::utils;
-
 pub mod config;
 pub mod cover_renderer;
-pub mod logging;
 pub mod patch_renderer;
 pub mod screens;
 
@@ -391,30 +389,30 @@ impl App {
     pub fn check_external_deps(&self) -> bool {
         let mut app_can_run = true;
 
-        if !utils::binary_exists("b4") {
+        if which::which("b4").is_err() {
             Logger::error("b4 is not installed, patchsets cannot be downloaded");
             app_can_run = false;
         }
 
-        if !utils::binary_exists("git") {
+        if which::which("git").is_err() {
             Logger::warn("git is not installed, send-email won't work");
         }
 
         match self.config.patch_renderer() {
             PatchRenderer::Bat => {
-                if !utils::binary_exists("bat") {
+                if which::which("bat").is_err() {
                     Logger::warn("bat is not installed, patch rendering will fallback to default");
                 }
             }
             PatchRenderer::Delta => {
-                if !utils::binary_exists("delta") {
+                if which::which("delta").is_err() {
                     Logger::warn(
                         "delta is not installed, patch rendering will fallback to default",
                     );
                 }
             }
             PatchRenderer::DiffSoFancy => {
-                if !utils::binary_exists("diff-so-fancy") {
+                if which::which("diff-so-fancy").is_err() {
                     Logger::warn(
                         "diff-so-fancy is not installed, patch rendering will fallback to default",
                     );
