+++ src/app/patch_renderer.rs
@@ -1,13 +1,13 @@
+use color_eyre::eyre::eyre;
+use serde::{Deserialize, Serialize};
+
 use std::{
     fmt::Display,
     io::Write,
     process::{Command, Stdio},
 };
 
-use color_eyre::eyre::eyre;
-use serde::{Deserialize, Serialize};
-
-use super::logging::Logger;
+use crate::infrastructure::logging::Logger;
 
 #[derive(Debug, Serialize, Deserialize, Clone, Copy, Default)]
 pub enum PatchRenderer {
