+++ src/infrastructure/logging/garbage_collector.rs
