+++ src/infrastructure/mod.rs
@@ -1,0 +1,3 @@
+pub mod errors;
+pub mod logging;
+pub mod terminal;
